+++ packages/exchange-consumer/src/index.ts
@@ -1,2 +1,3 @@
 export * from './consumers'
+export * from './streams'
 export * from './types'