+++ packages/spot-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.42](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.41...@injectivelabs/spot-consumer@0.0.42) (2021-06-13)
+
+
+### Bug Fixes
+
+* merge conflicts ([a5c7ce1](https://github.com/InjectiveLabs/injective-ts/commit/a5c7ce13823fded62a32f79fcfda19867c929cc7))
+
+
+
+
+
 ## [0.0.41](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.40...@injectivelabs/spot-consumer@0.0.41) (2021-06-08)
 
 **Note:** Version bump only for package @injectivelabs/spot-consumer